move another method up and linting
